Add switching modal form function
Added a function in the modal to switch
between project and to-do forms. Originally made
two separate buttons for adding a project and a
to-do, however I feel like putting them in the
same modal is more intuitive.

Edited HTML form so the input is wrapped in the
label element.

Added styles for the new wrapping containers
for both forms.